requirements.txt
Django==4.2.8
django-filter==23.5
django-guardian==2.4.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
requests-pkcs12==1.24
uvicorn==0.27.0.post1
python-dotenv